Reverting some commits for compatibility with Samsung's proprietary modules.
The two commits reverted below would cause the kernel to "oops" when using Samsung's proprietary modules.

Revert "PM / Wakeup: Introduce wakeup source objects and event statistics (v3)"
This reverts commit 599c6505ff2c07d52fcd92d4e81ee4f215d8b3fa.

Revert "mm: prevent concurrent unmap_mapping_range() on the same inode"
This reverts commit 3a9807556cf88598bf4cf182e0dd3fc659443bb4.

+++ Kernel/kernel/power/main.c
@@ -252,18 +252,18 @@ static ssize_t wakeup_count_show(struct kobject *kobj,
 				struct kobj_attribute *attr,
 				char *buf)
 {
-	unsigned int val;
+	unsigned long val;
 
-	return pm_get_wakeup_count(&val) ? sprintf(buf, "%u\n", val) : -EINTR;
+	return pm_get_wakeup_count(&val) ? sprintf(buf, "%lu\n", val) : -EINTR;
 }
 
 static ssize_t wakeup_count_store(struct kobject *kobj,
 				struct kobj_attribute *attr,
 				const char *buf, size_t n)
 {
-	unsigned int val;
+	unsigned long val;
 
-	if (sscanf(buf, "%u", &val) == 1) {
+	if (sscanf(buf, "%lu", &val) == 1) {
 		if (pm_save_wakeup_count(val))
 			return n;
 	}
